Remove 4-valued history
Displaying an average of the last 4 read values of a monitored resource
has little purpose. Additionally, it is questionable when initializing
the plugin if less than 4 values have been read so far (the number of
valid history values isn't being tracked anywhere).

The removed code is from year 2003 or before.