Add a simple network bandwidth monitor (uses libgtop as a fallback)
The implementation uses /proc/net/netstat by default (if it is available)
because it is more efficient than using libgtop. It takes about 250
syscalls for libgtop to obtain information about a couple of network
interfaces, while it takes just a few syscalls to read /proc/net/netstat.

A full network bar (100% network load) corresponds to a constant of
100 Mbit/s. Bandwidths higher than 100 Mbit/s are clamped to 100%.
The tooltip however is showing non-clamped bandwidth rounded to Mbit/s.

Closes: https://gitlab.xfce.org/panel-plugins/xfce4-systemload-plugin/-/issues/14